avgRating class
avgRating class to calculate the avg rating as well as store previous ratings in an array in backendless so we can continue to calculate the avg. ratings array created for meal and mealdata classes. ratings vars changed to doubles. avg rating was working in tableview. avg rating currently not working—succesfull merge with last push